ADD :: UserFacade 분리 (#24)
- getCurrentUser 빼려고 userFacade 만듦
- 근데 하는 김에 다뺐어염
