+++ src/main/java/com/bamdoliro/gati/domain/user/facade/UserFacade.java
@@ -1,0 +1,41 @@
+package com.bamdoliro.gati.domain.user.facade;
+
+import com.bamdoliro.gati.domain.user.domain.User;
+import com.bamdoliro.gati.domain.user.domain.repository.UserRepository;
+import com.bamdoliro.gati.domain.user.exception.PasswordMismatchException;
+import com.bamdoliro.gati.domain.user.exception.UserAlreadyExistsException;
+import com.bamdoliro.gati.domain.user.exception.UserNotFoundException;
+import com.bamdoliro.gati.global.security.auth.AuthDetails;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class UserFacade {
+
+    private final UserRepository userRepository;
+    private final PasswordEncoder passwordEncoder;
+
+    public User findUserByEmail(String email) {
+        return userRepository.findByEmail(email)
+                .orElseThrow(() -> UserNotFoundException.EXCEPTION);
+    }
+
+    public User getCurrentUser() {
+        AuthDetails auth = (AuthDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        return auth.getUser();
+    }
+
+    public void checkUserByEmail(String email) {
+        userRepository.findByEmail(email)
+                .ifPresent(user -> { throw UserAlreadyExistsException.EXCEPTION; });
+    }
+
+    public void checkUserPassword(User user, String password) {
+        if (!passwordEncoder.matches(password, user.getPassword())) {
+            throw PasswordMismatchException.EXCEPTION;
+        }
+    }
+}
